Refactor to a seperate files
