+++ dhcp-forwarder/forwarder.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"errors"
-	"net"
 
 	"github.com/google/gopacket"
 	"github.com/google/gopacket/pcap"
@@ -15,61 +14,6 @@ type ForwardHandler interface {
 type Forwarder struct {
 	Handler ForwardHandler
 	BPF     *pcap.BPF
-}
-
-type SimpleUDPHandler struct {
-	conn *net.UDPConn
-}
-
-func (h *SimpleUDPHandler) Forward(p gopacket.Packet) error {
-	udpLayer := p.TransportLayer()
-	if udpLayer != nil {
-		conn.Write(udpLayer.LayerPayload())
-		// We don't check for error here.
-		// The endpoint might not be listening yet.
-	}
-
-	if errLayer := p.ErrorLayer(); errLayer != nil {
-		return errLayer.Error()
-	}
-
-	return nil
-}
-
-func NewSimpleUDPHandler(c *ForwarderConfig) (ForwardHandler, error) {
-	udpAddr, err := net.ResolveUDPAddr("udp4", c.Host+":"+c.Port)
-	if err != nil {
-		return nil, err
-	}
-
-	conn, err = net.DialUDP("udp", nil, udpAddr)
-	if err != nil {
-		return nil, err
-	}
-
-	return &SimpleUDPHandler{conn: conn}, nil
-}
-
-type DNSHandler struct {
-	conn *net.UDPConn
-}
-
-func NewDNSHandler(c *ForwarderConfig) (ForwardHandler, error) {
-	udpAddr, err := net.ResolveUDPAddr("udp4", c.Host+":"+c.Port)
-	if err != nil {
-		return nil, err
-	}
-
-	conn, err = net.DialUDP("udp", nil, udpAddr)
-	if err != nil {
-		return nil, err
-	}
-
-	return &DNSHandler{conn: conn}, nil
-}
-
-func (h *DNSHandler) Forward(p gopacket.Packet) error {
-	return nil
 }
 
 type ForwardHandlerBuilder func(c *ForwarderConfig) (ForwardHandler, error)
